tests/submission/W3C/domhtml/HTMLModElement01.html

<!DOCTYPE html>
<html><head>
<meta charset="utf-8">
<title>NIST DOM HTML Test - MOD</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="../common/domtests-common.js"></script>
<link href="http://www.w3.org/TR/html5/attributes-common-to-ins-and-del-elements.html#dom-mod-cite" rel="help">
<link rel="author" title="NIST">
</head>
<body>
<p>
<ins cite="./files/ins-reasons.html" datetime="January 1, 2002">The INS element is used to indicate that a section of a document had been inserted.</ins>
<br>
<del cite="./files/del-reasons.html" datetime="January 2, 2002">The DEL element is used to indicate that a section of a document had been removed.</del>
</p>



<script>
var domtest = async_test("The cite attribute specifies an URI designating a document that describes     the reason for the change.");

on_event(window, "load", domtest.step_func(function (dom_event) {
    var nodeList;
      var testNode;
      var vcite;
      var doc;
      doc = window.document;
      nodeList = doc.getElementsByTagName("ins");
      assert_equals(nodeList.length,1);
testNode = nodeList.item(0);
      vcite = testNode.cite;

      assert_uri_equals(null,null,null,"ins-reasons.html",null,null,null,null,vcite);

    domtest.done();
}));

</script>
<p>The cite attribute specifies an URI designating a document that describes
    the reason for the change.

    Retrieve the cite attribute of the INS element and examine its value.</p>
<div id="log"></div>
</body></html>
